refactor: reorganize imports and streamline node registration in audio transcription
